Project Domain integrated test

+++ src/main/java/org/certis/studyplatform/member/application/object/command/UpdatePenaltyCommand.java
@@ -1,8 +1,8 @@
 package org.certis.studyplatform.member.application.object.command;
 
 public record UpdatePenaltyCommand( Long memberId,
-                                    Long penaltyPoints) {
-    public static UpdatePenaltyCommand of(Long memberId, Long penaltyPoints) {
+                                    Integer penaltyPoints) {
+    public static UpdatePenaltyCommand of(Long memberId, Integer penaltyPoints) {
         return new UpdatePenaltyCommand(memberId, penaltyPoints);
     }
 }
